+++ lib/SimpleHttpClient/GeneralClientConnection.h
@@ -71,7 +71,7 @@ class GeneralClientConnection {
   //////////////////////////////////////////////////////////////////////////////
 
   static GeneralClientConnection* factory(Endpoint*, double,
-                                          double, size_t, uint32_t);
+                                          double, size_t, uint64_t);
 
   static GeneralClientConnection* factory(
       std::unique_ptr<Endpoint>&, double, double, size_t,
